+++ Module.php
@@ -8,6 +8,7 @@ use humhub\modules\mail\permissions\StartConversation;
 use humhub\modules\mail\permissions\SendMail;
 use humhub\modules\user\models\User;
 use humhub\modules\mail\helpers\Url;
+use Yii;
 
 /**
  * MailModule provides messaging functions inside the application.
@@ -22,6 +23,34 @@ class Module extends \humhub\components\Module
      * @inheritdoc
      */
     public $resourcesPath = 'resources';
+
+    /**
+     * @var int Defines the initial page size of conversation overview inbox sidebar in messenger view. Should be > 6
+     */
+    public $inboxInitPageSize = 30;
+
+    /**
+     * @var int Defines the page size when loading more conversation page entries (autoscroller)
+     */
+    public $inboxUpdatePageSize = 5;
+
+    /**
+     * @var int Defines the initial message amount loaded for a conversation
+     */
+    public $conversationInitPageSize = 50;
+
+    /**
+     * @var int Defines the amount of messages loaded when loading more messages
+     */
+    public $conversationUpdatePageSize = 50;
+
+    /**
+     * @return static
+     */
+    public static function getModuleInstance()
+    {
+        return Yii::$app->getModule('mail');
+    }
 
     /**
      * @inheritdoc