+++ resources/js/humhub.mail.inbox.js
@@ -2,6 +2,9 @@ humhub.module('mail.inbox', function (module, require, $) {
 
     var Widget = require('ui.widget').Widget;
     var Filter = require('ui.filter').Filter;
+    var view = require('ui.view');
+    var loader = require('ui.loader');
+    var client = require('client');
 
     var ConversationFilter = Filter.extend();
 
@@ -31,18 +34,131 @@ humhub.module('mail.inbox', function (module, require, $) {
     ConversationList.prototype.init = function () {
         this.filter = Widget.instance('#mail-filter-root');
 
+        this.initScroll();
+
         var that = this;
         this.filter.off('afterChange.inbox').on('afterChange.inbox', function () {
-            that.reload();
+            that.reload().then(function() {
+                that.updateActiveItem();
+            });
         });
+
+        if(view.isLarge()) {
+            this.$.niceScroll({
+                cursorwidth: "7",
+                cursorborder: "",
+                cursorcolor: "#555",
+                cursoropacitymax: "0.2",
+                nativeparentscrolling: false,
+                railpadding: {top: 0, right: 3, left: 0, bottom: 0}
+            });
+        }
+    };
+
+    ConversationList.prototype.initScroll = function() {
+        if (window.IntersectionObserver) {
+
+            var $streamEnd = $('<div class="stream-end"></div>');
+            this.$.append($streamEnd);
+
+            var that = this;
+            var observer = new IntersectionObserver(function (entries) {
+                if (that.preventScrollLoading()) {
+                    return;
+                }
+
+                if (entries.length && entries[0].isIntersecting) {
+                    loader.append(that.$);
+                    that.loadMore().finally(function() {
+                        loader.reset(that.$);
+                    });
+                }
+
+            }, {root: this.$[0], rootMargin: "50px"});
+
+            // Assure the conversation list is scrollable by loading more entries until overflow
+            this.assureScroll().then(function() {
+                observer.observe($streamEnd[0]);
+            });
+        }
+    };
+
+    ConversationList.prototype.assureScroll = function () {
+        var that = this;
+
+        if(this.$[0].offsetHeight >= this.$[0].scrollHeight && this.canLoadMore()) {
+            return this.loadMore().then(function() {
+                return that.assureScroll();
+            }).catch(function () {
+                return Promise.resolve();
+            })
+        }
+
+        return Promise.resolve();
+    };
+
+    ConversationList.prototype.loadMore = function () {
+        var that = this;
+        return new Promise(function(resolve, reject) {
+            var data = that.filter.getFilterMap();
+            data.from = that.getLastMessageId();
+            client.get(that.options.loadMoreUrl, {data: data}).then(function(response) {
+                if(response.result) {
+                    $(response.result).insertBefore('.stream-end');
+                    that.$.find('.stream-end').append();
+                }
+
+                that.options.isLast = !response.result || response.isLast;
+                that.updateActiveItem();
+
+                resolve();
+            }).catch(function(err) {
+                module.log.error(err, true);
+                reject();
+            }).finally(function() {
+                that.scrollLock = false;
+            });
+        });
+
+    };
+
+    ConversationList.prototype.preventScrollLoading = function () {
+        return this.scrollLock || !this.canLoadMore();
+    };
+
+    ConversationList.prototype.canLoadMore = function () {
+        return !this.options.isLast;
     };
 
     ConversationList.prototype.getReloadOptions = function () {
         return {data: this.filter.getFilterMap()};
     };
 
+    ConversationList.prototype.updateActiveItem = function() {
+
+        var activeMessageId = Widget.instance('#mail-conversation-root').getActiveMessageId();
+
+        this.$.find('.entry').removeClass('selected');
+
+        // Remove New badge from current selection
+        this.$.find('.entry.selected').find('.new-message-badge').hide();
+
+        // Set new selection
+        this.$.find('.entry').removeClass('selected');
+        var $selected = this.$.find('[data-message-preview="' + activeMessageId + '"]');
+
+        if($selected.length) {
+            $selected.addClass('selected').find('.new-message-badge').hide();
+        }
+    };
+
+
     ConversationList.prototype.getFirstMessageId = function() {
-        return this.$.find('.messagePreviewEntry:first').data('messagePreview');
+        return this.$.find('.entry:first').data('messagePreview');
+    };
+
+    ConversationList.prototype.getLastMessageId = function() {
+        return this.$.find('.entry:last').data('messagePreview');
     };
 
     var setTagFilter = function (evt) {