+++ views/mail/conversation.php
@@ -2,7 +2,9 @@
 
 use humhub\libs\Html;
 use humhub\modules\mail\models\forms\ReplyForm;
+use humhub\modules\mail\widgets\ConversationHeader;
 use humhub\modules\mail\widgets\ConversationTags;
+use humhub\modules\mail\widgets\Messages;
 use humhub\modules\ui\view\components\View;
 use yii\bootstrap\ActiveForm;
 use humhub\modules\content\widgets\richtext\ProsemirrorRichTextEditor;
@@ -25,7 +27,7 @@ use humhub\widgets\Button;
     <?php  else :?>
 
         <div id="mail-conversation-header" class="panel-heading" style="background-color:<?= $this->theme->variable('background-color-secondary')?>">
-            <strong> <?= $this->render('_conversation_header', ['message' => $message]) ?></strong>
+            <strong> <?= ConversationHeader::widget(['message' => $message]) ?></strong>
         </div>
 
         <?= ConversationTags::widget(['message' => $message])?>
@@ -33,12 +35,7 @@ use humhub\widgets\Button;
         <div class="panel-body">
 
             <div class="media-list conversation-entry-list">
-                <?php $prevEntry = null; ?>
-                <?php foreach ($message->entries as $index => $entry) : ?>
-                    <?php $nextEntry = $message->entries[$index + 1] ?? null ?>
-                    <?= ConversationEntry::widget(['entry' => $entry, 'prevEntry' => $prevEntry, 'nextEntry' => $nextEntry])?>
-                    <?php $prevEntry = $entry ?>
-                <?php endforeach; ?>
+                <?= Messages::widget(['message' => $message])?>
             </div>
 
 
@@ -49,7 +46,7 @@ use humhub\widgets\Button;
                         ProsemirrorRichTextEditor::class, [
                         'menuClass' => 'plainMenu',
                         'placeholder' => Yii::t('MailModule.base', 'Write a message...'),
-                        'pluginOptions' => ['maxHeight' => '300px'],
+                        'pluginOptions' => ['maxHeight' => '200px'],
                     ])->label(false) ?>
 
                     <?= Button::primary(Yii::t('MailModule.views_mail_show', 'Send'))->submit()->action('reply', $replyForm->getUrl()) ?>