+++ widgets/ConversationInbox.php
@@ -6,6 +6,8 @@ namespace humhub\modules\mail\widgets;
 
 use humhub\modules\mail\helpers\Url;
 use humhub\modules\mail\models\forms\InboxFilterForm;
+use humhub\modules\mail\models\UserMessage;
+use humhub\modules\mail\Module;
 use humhub\widgets\JsWidget;
 
 class ConversationInbox extends JsWidget
@@ -33,25 +35,36 @@ class ConversationInbox extends JsWidget
     /**
      * @inheritDoc
      */
-    public $pageSize = 30;
+    public $pageSize = 7;
+
+    /**
+     * @var UserMessage[]
+     */
+    private $result;
+
+    public function init()
+    {
+        parent::init();
+        $this->result = $this->filter->getPage();
+    }
 
     /**
      * @inheritDoc
      */
     public function run()
     {
-        $this->filter->apply();
-
         return $this->render('inbox', [
             'options' => $this->getOptions(),
-            'userMessages' =>  $this->filter->query->limit($this->pageSize)->all(),
+            'userMessages' =>  $this->result
         ]);
     }
 
     public function getData()
     {
         return [
-            'widget-reload-url' => Url::toUpdateInbox()
+            'widget-reload-url' => Url::toUpdateInbox(),
+            'load-more-url' => Url::toInboxLoadMore(),
+            'is-last' => $this->filter->wasLastPage()
         ];
     }